Add ConversationPolicy and conversation update fix.

+++ app/Http/Controllers/ConversationController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\Conversation\StoreConversationRequest;
 use App\Http\Requests\Conversation\UpdateConversationRequest;
 use App\Http\Resources\Conversation\ConversationResource;
 use App\Models\Conversation;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
@@ -15,9 +16,12 @@ class ConversationController extends Controller
 {
     /**
      * @return AnonymousResourceCollection
+     * @throws AuthorizationException
      */
     public function index(): AnonymousResourceCollection
     {
+        $this->authorize('viewAny', Conversation::class);
+
         $conversations = Conversation::with('latestMessage')
             ->withReceiverUsers()
             ->orderBy('created_at', 'DESC')
@@ -29,9 +33,12 @@ class ConversationController extends Controller
     /**
      * @param StoreConversationRequest $request
      * @return ConversationResource
+     * @throws AuthorizationException
      */
     public function store(StoreConversationRequest $request): ConversationResource
     {
+        $this->authorize('create', Conversation::class);
+
         DB::beginTransaction();
         try {
             $conversation = Conversation::create($request->validated());
@@ -50,10 +57,13 @@ class ConversationController extends Controller
     /**
      * @param Conversation $conversation
      * @return ConversationResource
+     * @throws AuthorizationException
      */
     public function show(Conversation $conversation): ConversationResource
     {
-        $conversation = Conversation::where('id', $conversation->id)->with(['messages', 'users'])->get();
+        $this->authorize('view', $conversation);
+
+        $conversation = Conversation::where('id', $conversation->id)->with(['messages', 'users'])->first();
 
         return ConversationResource::make($conversation);
     }
@@ -62,10 +72,23 @@ class ConversationController extends Controller
      * @param UpdateConversationRequest $request
      * @param Conversation $conversation
      * @return ConversationResource
+     * @throws AuthorizationException
      */
     public function update(UpdateConversationRequest $request, Conversation $conversation): ConversationResource
     {
-        $conversation->update($request->validated());
+        $this->authorize('update', $conversation);
+
+        DB::beginTransaction();
+        try {
+            $conversation->update($request->validated());
+            if ($request->has('to_user_id')) {
+                $conversation->users()->syncWithoutDetaching($request->to_user_id);
+            }
+        } catch (Exception $exception) {
+            DB::rollBack();
+            abort(500, 'Somethings wrong');
+        }
+        DB::commit();
 
         return ConversationResource::make($conversation);
     }
@@ -73,9 +96,12 @@ class ConversationController extends Controller
     /**
      * @param Conversation $conversation
      * @return Response
+     * @throws AuthorizationException
      */
     public function destroy(Conversation $conversation): Response
     {
+        $this->authorize('delete', $conversation);
+
         try {
             $conversation->delete();
         } catch (\Exception $exception) {
